Update GitHub actions, upload-pages-artifact and deploy-pages.
The main change deploy-pages v2 brings over v1 are dependency updates.

The main change upload-pages-artifact v2 brings over v1 is the
removal of the chmod command built into the action. This was slowing
down deployment for large user and causing issues where it was setting
the permission incorrectly on certain files. The recommendation is tom
manually set the permission yourself when needed.