Upgrade blog to use jekyll 4.3.0 from 3.9.2.
The theme used (minima) needs to to be updated soon when it stops using
a deprecrated feature. For the time, being the sass converter has been
downgraded so it d oesn't issue the deprecration warning as its a known
issue with minima.

The jekyll-build-pages action can not be used as it only supports
jekyll 3.x and they have not released a new version that uses 4.